% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{argsstring2list}
\alias{argsstring2list}
\title{Convert string of arguments to list of arguments}
\usage{
argsstring2list(string)
}
\arguments{
\item{string}{string to convert to list to use with do.call}
}
\value{
list
}
\description{
Idea from the answer: https://stackoverflow.com/a/62979238
}
