% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_table.R
\name{baseline_table}
\alias{baseline_table}
\title{Print a flexible baseline characteristics table}
\usage{
baseline_table(
  data,
  fun.args = NULL,
  fun = gtsummary::tbl_summary,
  vars = NULL
)
}
\arguments{
\item{data}{data set}

\item{fun.args}{list of arguments passed to}

\item{fun}{function to}

\item{vars}{character vector of variables to include}
}
\value{
object of standard class for fun
}
\description{
Print a flexible baseline characteristics table
}
\examples{
mtcars |> baseline_table()
mtcars |> baseline_table(fun.args = list(by = "gear"))
}
