% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{default_parsing}
\alias{default_parsing}
\title{Default data parsing}
\usage{
default_parsing(data)
}
\arguments{
\item{data}{data}
}
\value{
data.frame or tibble
}
\description{
Default data parsing
}
\examples{
mtcars |> str()
mtcars |>
  default_parsing() |>
  str()
}
