% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{factorize}
\alias{factorize}
\title{Factorize variables in data.frame}
\usage{
factorize(data, vars)
}
\arguments{
\item{data}{data.frame}

\item{vars}{variables to force factorize}
}
\value{
data.frame
}
\description{
Factorize variables in data.frame
}
