% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{file_export}
\alias{file_export}
\title{Title}
\usage{
file_export(data, output.format = c("df", "teal", "list"), filename, ...)
}
\arguments{
\item{data}{data}

\item{output.format}{output}

\item{filename}{filename}

\item{...}{passed on}
}
\value{
data
}
\description{
Title
}
