% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{format_writer}
\alias{format_writer}
\title{Merges list of named arguments for qmd header generation}
\usage{
format_writer(data, name)
}
\arguments{
\item{data}{vector}

\item{name}{name}
}
\value{
vector
}
\description{
Merges list of named arguments for qmd header generation
}
