% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getfun}
\alias{getfun}
\title{Wrapper function to get function from character vector referring to function from namespace. Passed to 'do.call()'}
\usage{
getfun(x)
}
\arguments{
\item{x}{function or function name}
}
\value{
function or character vector
}
\description{
This function follows the idea from this comment: https://stackoverflow.com/questions/38983179/do-call-a-function-in-r-without-loading-the-package
}
\examples{
getfun("stats::lm")
}
