% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{index_embed}
\alias{index_embed}
\title{Split vector by an index and embed addition}
\usage{
index_embed(data, index, add = NULL)
}
\arguments{
\item{data}{vector}

\item{index}{split index}

\item{add}{addition}
}
\value{
vector
}
\description{
Split vector by an index and embed addition
}
