% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-variable-dates.R
\name{is_any_class}
\alias{is_any_class}
\title{Test class}
\usage{
is_any_class(data, class.vec)
}
\arguments{
\item{data}{data}

\item{class.vec}{vector of class names to test}
}
\value{
factor
}
\description{
Test class
}
\examples{
vapply(REDCapCAST::redcapcast_data, \(.x){
  is_any_class(.x, c("hms", "Date", "POSIXct", "POSIXt"))
}, logical(1))
}
