% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{modify_qmd}
\alias{modify_qmd}
\title{Wrapper to modify quarto file to render specific formats}
\usage{
modify_qmd(file, format)
}
\arguments{
\item{file}{filename}

\item{format}{desired output}
}
\value{
none
}
\description{
Wrapper to modify quarto file to render specific formats
}
