% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{panel_space}
\alias{panel_space}
\title{Wrapping nav_spacer to avoid errors on dependencies when packaging}
\usage{
panel_space()
}
\value{
bslib object
}
\description{
Wrapping nav_spacer to avoid errors on dependencies when packaging
}
