% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_read_shiny_module.R
\docType{data}
\name{m_redcap_readUI}
\alias{m_redcap_readUI}
\alias{m_redcap_readServer}
\alias{tdm_redcap_read}
\alias{redcap_app}
\title{Shiny module to browser and export REDCap data}
\format{
An object of class \code{teal_data_module} of length 2.
}
\usage{
m_redcap_readUI(id, include_title = TRUE)

m_redcap_readServer(id, output.format = c("df", "teal", "list"))

tdm_redcap_read

redcap_app()
}
\arguments{
\item{id}{Namespace id}

\item{include_title}{logical to include title}

\item{output.format}{data.frame ("df") or teal data object ("teal")}
}
\value{
shiny ui element

shiny server module
}
\description{
Shiny module to browser and export REDCap data

REDCap import teal data module

Test app for the redcap_read_shiny_module
}
\examples{
\dontrun{
redcap_app()
}
}
\keyword{datasets}
