% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shiny_webResearch}
\alias{shiny_webResearch}
\title{Test version of the shiny_cast function to launch the app with a data set in
the environment.}
\usage{
shiny_webResearch(data = NULL, ...)
}
\arguments{
\item{data}{optional data set to provide for analysis}

\item{...}{arguments passed on to \code{shiny::runApp()}}
}
\value{
shiny app
}
\description{
Test version of the shiny_cast function to launch the app with a data set in
the environment.
}
\examples{
\dontrun{
mtcars |> shiny_webResearch(launch.browser = TRUE)
}
}
