% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{specify_qmd_format}
\alias{specify_qmd_format}
\title{Specify format arguments to include in qmd header/frontmatter}
\usage{
specify_qmd_format(data, fileformat = c("docx", "odt", "pdf", "all"))
}
\arguments{
\item{data}{vector}

\item{fileformat}{format to include}
}
\value{
vector
}
\description{
Specify format arguments to include in qmd header/frontmatter
}
