% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_table.R
\name{tbl_merge}
\alias{tbl_merge}
\title{A substitue to gtsummary::tbl_merge, that will use list names for the tab
spanner names.}
\usage{
tbl_merge(data)
}
\arguments{
\item{data}{gtsummary list object}
}
\value{
gt summary list object
}
\description{
A substitue to gtsummary::tbl_merge, that will use list names for the tab
spanner names.
}
