% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{m_redcap_readUI}
\alias{m_redcap_readUI}
\alias{m_redcap_readServer}
\title{Shiny module to browser and export REDCap data}
\usage{
m_redcap_readUI(id)

m_redcap_readServer(id, output.format = "df")
}
\arguments{
\item{id}{Namespace id}

\item{output.format}{data.frame ("df") or teal data object ("teal")}
}
\value{
shiny ui element

shiny server module
}
\description{
Shiny module to browser and export REDCap data
}
