% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_quarto}
\alias{write_quarto}
\title{Wrapper to save data in RDS, load into specified qmd and render}
\usage{
write_quarto(
  data,
  fileformat,
  qmd.file = here::here("analyses.qmd"),
  file = NULL,
  ...
)
}
\arguments{
\item{data}{list to pass to qmd}

\item{fileformat}{output format. Ignored if file!=NULL}

\item{qmd.file}{qmd file to render. Default is 'here::here("analyses.qmd")'}

\item{file}{exact filename (Optional)}

\item{...}{Ignored for now}
}
\value{
none
}
\description{
Wrapper to save data in RDS, load into specified qmd and render
}
